python-agents/requirements.txt
crewai==0.41.1
crewai-tools==0.8.3
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6
cors==1.0.1
